spaces, and their classes, moved to different folder
