--- src/main/java/ch/uzh/ifi/hase/soprafs24/gamelogic/Card.java
@@ -1,4 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.gamelogic;
-
-public class Card {
-}
