+++ src/main/java/ch/uzh/ifi/hase/soprafs24/logic/Game/BlackBigOops.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.gamelogic;
+package ch.uzh.ifi.hase.soprafs24.logic.Game;
 
 public class BlackBigOops {
     private static final String[] blacks = {"swapPosiOfAllPlayers", "lose69Cash", "evenlyDistributeAllCash"};
