+++ src/main/java/ch/uzh/ifi/hase/soprafs24/logic/Game/Card.java
@@ -1,0 +1,4 @@
+package ch.uzh.ifi.hase.soprafs24.logic.Game;
+
+public class Card {
+}
