+++ src/main/java/ch/uzh/ifi/hase/soprafs24/logic/Game/Player.java
@@ -1,16 +1,101 @@
 package ch.uzh.ifi.hase.soprafs24.logic.Game; //NOSONAR
 
 
+import java.util.List;
+
 public class Player {
 
-    private int id; //NOSONAR
-    private String name; //NOSONAR
-    private int cash; //NOSONAR
-    //TODO verify how position is saved for player in backend (id of space or space itself) NOSONAR
-    private int position; //NOSONAR
-    //TODO import card, item and ultimate NOSONAR
-    //ArrayList<Card> cards; //NOSONAR
-    //ArrayList<Item> items; //NOSONAR
-    //private Ultimate ultimate; //NOSONAR
-    private Player teammateId; //NOSONAR
+    private Long playerId;
+    private String playerName;
+    private int cash;
+    private Long position;
+
+    private List<Card> cards;
+    private List<Item> items;
+    private Ultimate ultimate;
+
+    private Long userId;
+    private Long teammateId;
+    private Long gameBoardId;
+
+    public Long getPlayerId() {
+        return playerId;
+    }
+
+    public void setPlayerId(Long playerId) {
+        this.playerId = playerId;
+    }
+
+    public String getPlayerName() {
+        return playerName;
+    }
+
+    public void setPlayerName(String playerName) {
+        this.playerName = playerName;
+    }
+
+    public int getCash() {
+        return cash;
+    }
+
+    public void setCash(int cash) {
+        this.cash = cash;
+    }
+
+    public Long getPosition() {
+        return position;
+    }
+
+    public void setPosition(Long position) {
+        this.position = position;
+    }
+
+    public List<Card> getCards() {
+        return cards;
+    }
+
+    public void setCards(List<Card> cards) {
+        this.cards = cards;
+    }
+
+    public List<Item> getItems() {
+        return items;
+    }
+
+    public void setItems(List<Item> items) {
+        this.items = items;
+    }
+
+    public Ultimate getUltimate() {
+        return ultimate;
+    }
+
+    public void setUltimate(Ultimate ultimate) {
+        this.ultimate = ultimate;
+    }
+
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Long userId) {
+        this.userId = userId;
+    }
+
+    public Long getTeammateId() {
+        return teammateId;
+    }
+
+    public void setTeammateId(Long teammateId) {
+        this.teammateId = teammateId;
+    }
+
+    public Long getGameBoardId() {
+        return gameBoardId;
+    }
+
+    public void setGameBoardId(Long gameBoardId) {
+        this.gameBoardId = gameBoardId;
+    }
 }
