+++ src/main/java/ch/uzh/ifi/hase/soprafs24/logic/Game/RedSmallOops.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.gamelogic;
+package ch.uzh.ifi.hase.soprafs24.logic.Game;
 
 public class RedSmallOops {
     private static final String[] reds = {"youLose10Cash", "everyoneLose10Cash", "tpToStart"};
