+++ src/main/java/ch/uzh/ifi/hase/soprafs24/logic/Game/WinCondi.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.gamelogic;
+package ch.uzh.ifi.hase.soprafs24.logic.Game;
 
 public class WinCondi {
     private static final String[] winCondis = {"JackSparrow", "theMarooned", "goldenIsMy...", "drunk"};
